Fixed the glitch in Orange animation
The glitch was that when Orange turned from platform
into player, after the transition was complete,
there was a single frame where he would become a platform
again, and then player (and stay as player).

The problem was the the transition from "platform to human"
animation to "idle human" animation was taking some time,
and during this time the "platform->human" animation started
playing from the beginning.